Improved Software Design - Reusability by making all nested-preferences use the same Activity and PreferenceFragment. Added Support for 3NOW application. Removed references of JSoup lib as its not needed (reducing file size)

+++ app/src/main/java/nz/pbomb/xposed/anzmods/PrefFragment.java
@@ -1,5 +1,6 @@
 package nz.pbomb.xposed.anzmods;
 
+import android.annotation.SuppressLint;
 import android.app.ActivityManager;
 import android.app.AlertDialog;
 import android.content.Context;
@@ -9,6 +10,7 @@ import android.os.Bundle;
 import android.preference.CheckBoxPreference;
 import android.preference.Preference;
 import android.preference.PreferenceFragment;
+import android.util.Log;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -19,58 +21,60 @@ import common.PACKAGES;
 import common.PREFERENCES;
 
 @SuppressWarnings("unchecked")
-public class ANZPrefFragment extends PreferenceFragment implements Preference.OnPreferenceChangeListener {
-    private SharedPreferences sharedPreferences;
-    private Map<String, Boolean> oldPreferences;
+public class PrefFragment extends PreferenceFragment implements SharedPreferences.OnSharedPreferenceChangeListener, Preference.OnPreferenceChangeListener {
+    private Map<String, ?> oldPreferences;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        addPreferencesFromResource(getArguments().getInt("id"));
 
-        addPreferencesFromResource(R.xml.preferences_anz);
-        this.sharedPreferences = getActivity().getSharedPreferences(PREFERENCES.SHARED_PREFS_FILE_NAME, Context.MODE_WORLD_READABLE);
-        oldPreferences = (Map<String, Boolean>) sharedPreferences.getAll();
+        //this.sharedPreferences = getActivity().getSharedPreferences(PREFERENCES.SHARED_PREFS_FILE_NAME, Context.MODE_WORLD_READABLE);
+        //oldPreferences = (Map<String, Boolean>) sharedPreferences.getAll();
+        getPreferenceManager().setSharedPreferencesMode(Context.MODE_WORLD_READABLE);
+        getPreferenceManager().getSharedPreferences().registerOnSharedPreferenceChangeListener(this);
+        oldPreferences = getPreferenceManager().getSharedPreferences().getAll();
 
+        if(getArguments() != null) {
+            if(getArguments().containsKey("preference")) {
+                if (getArguments().getString("preference").equals(PREFERENCES.KEYS.MAIN.ANZ)) {
+                    //getPreferenceManager().findPreference(PREFERENCES.KEYS.ANZ.ROOT_DETECTION).setOnPreferenceChangeListener(this);
+                    getPreferenceManager().findPreference(PREFERENCES.KEYS.ANZ.SPOOF_DEVICE).setOnPreferenceChangeListener(this);
+                    getPreferenceManager().findPreference(PREFERENCES.KEYS.ANZ.SCREENSHOT_ENABLED).setOnPreferenceChangeListener(this);
+                }
+            }
+        }
+    }
 
-        getPreferenceManager().findPreference(PREFERENCES.KEYS.ANZ.ROOT_DETECTION).setOnPreferenceChangeListener(this);
-        getPreferenceManager().findPreference(PREFERENCES.KEYS.ANZ.SPOOF_DEVICE).setOnPreferenceChangeListener(this);
-        getPreferenceManager().findPreference(PREFERENCES.KEYS.ANZ.SCREENSHOT_ENABLED).setOnPreferenceChangeListener(this);
+    @Override
+    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
     }
 
     @Override
     public boolean onPreferenceChange(Preference preference, Object newValue) {
-        String key = preference.getKey();
-        SharedPreferences.Editor sharedPreferencesEditor = sharedPreferences.edit();
+        CheckBoxPreference cp = ((CheckBoxPreference) preference);
+        String key = cp.getKey();
 
-        if (!((CheckBoxPreference) preference).isChecked()) {
-            sharedPreferencesEditor.putBoolean(key, true).apply();
+        if(getArguments().getString("preference").equals(PREFERENCES.KEYS.MAIN.ANZ)) {
+            if(key.equals(PREFERENCES.KEYS.ANZ.SPOOF_DEVICE)) {
+                if (Boolean.valueOf(newValue.toString())) {
+                    displaySpoofDeviceCheckedDialog();
+                }
+            } else if(key.equals(PREFERENCES.KEYS.ANZ.SCREENSHOT_ENABLED)) {
+                // The primary way of killing the ANZ GoMoney application
+                ActivityManager mActivityManager = (ActivityManager) getActivity().getApplicationContext().getSystemService(Context.ACTIVITY_SERVICE);
+                mActivityManager.killBackgroundProcesses(PACKAGES.ANZ_GOMONEY);
 
-            /*if(key.equals(PREFERENCES.KEYS.ANZ.SCREENSHOT_ENABLED)) {
-                displayScreenshotsEnabledCheckedDialog();
-            } else*/ if(key.equals(PREFERENCES.KEYS.ANZ.SPOOF_DEVICE)) {
-                displaySpoofDeviceCheckedAlertDialog();
-            }
-
-        } else {
-            sharedPreferencesEditor.putBoolean(key, false).apply();
-        }
-
-        // If the Screenshot feature was toggled, we need to kill any ANZ processes in order for the setting to be in effect
-        if(key.equals(PREFERENCES.KEYS.ANZ.SCREENSHOT_ENABLED)) {
-            // The primary way of killing the ANZ GoMoney application
-            ActivityManager mActivityManager = (ActivityManager) getActivity().getApplicationContext().getSystemService(Context.ACTIVITY_SERVICE);
-            mActivityManager.killBackgroundProcesses(PACKAGES.ANZ_GOMONEY);
-
-            // The secondary way of killing the ANZ GoMoney application used as a backup.
-            // This is mainly for lollipop and TouchWiz ROMs where the primary way may not always work.
-            try {
-                Process process = Runtime.getRuntime().exec(new String[] { "am force-stop " + PACKAGES.ANZ_GOMONEY });
-                BufferedReader in = new BufferedReader(new InputStreamReader(process.getInputStream()));
-            } catch (IOException e) {
-                //e.printStackTrace();
+                // The secondary way of killing the ANZ GoMoney application used as a backup.
+                // This is mainly for lollipop and TouchWiz ROMs where the primary way may not always work.
+                try {
+                    Process process = Runtime.getRuntime().exec(new String[] { "am force-stop " + PACKAGES.ANZ_GOMONEY });
+                    BufferedReader in = new BufferedReader(new InputStreamReader(process.getInputStream()));
+                } catch (IOException e) {
+                    //e.printStackTrace();
+                }
             }
         }
-
         return true;
     }
 
@@ -80,7 +84,7 @@ public class ANZPrefFragment extends PreferenceFragment implements Preference.On
      */
     public boolean hasValuesChanged() {
         // Retrieve the latest SharedPreferences
-        Map<String, Boolean> currentPreferences = (Map<String, Boolean>) sharedPreferences.getAll();
+        Map<String, ?> currentPreferences = getPreferenceManager().getSharedPreferences().getAll();
         // Loop through all keys and compare the old preferences with the new preferences to detect
         // a change
         for(String key: oldPreferences.keySet()) {
@@ -101,7 +105,7 @@ public class ANZPrefFragment extends PreferenceFragment implements Preference.On
     /**
      * Displays the Alert Dialog when leaving the application
      */
-    public void displaySpoofDeviceCheckedAlertDialog() {
+    public void displaySpoofDeviceCheckedDialog() {
         AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
         builder.setMessage(getResources().getString(R.string.ANZPrefActivity_spoofDeviceChecked_message));
         builder.setCancelable(false);
@@ -114,6 +118,7 @@ public class ANZPrefFragment extends PreferenceFragment implements Preference.On
         AlertDialog alert = builder.create();
         alert.show();
     }
+
 
     /**
      * Displays the Alert Dialog when leaving the application
